refactor(codegen): parse introspection XML through one shared config
The xmlutil parser configuration existed in three identical copies: the
shipped CLI (Xml2Kotlin), TestXmlSupport, and GeneratorTest. The golden
fixture tests therefore asserted generator output against a parser they
configured themselves, so the suite could stay green through a regression
in the parsing path the tool actually ships.

Extract it once as `internal val introspectionXml` in the main source set
and point all three call sites at it. Behaviour is unchanged: the 66
checked-in fixture files are byte-identical.

Also documents/suppresses the xmlutil 1.0 `XML.compat` deprecation at the
now-single site — its replacements deliberately change the parsing
defaults and xmlutil exposes no non-deprecated route to the old ones, so
migrating is a behaviour change that needs its own review.

Fixes #166
