+++ codegen/src/main/kotlin/Xml2Kotlin.kt
@@ -80,14 +80,6 @@ import com.github.ajalt.clikt.parameters.types.file
 import java.io.File
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.decodeFromString
-import nl.adaptivity.xmlutil.ExperimentalXmlUtilApi
-import nl.adaptivity.xmlutil.QName
-import nl.adaptivity.xmlutil.XmlReader
-import nl.adaptivity.xmlutil.serialization.DefaultXmlSerializationPolicy
-import nl.adaptivity.xmlutil.serialization.InputKind
-import nl.adaptivity.xmlutil.serialization.XML
-import nl.adaptivity.xmlutil.serialization.XML.ParsedData
-import nl.adaptivity.xmlutil.serialization.structure.XmlDescriptor
 
 fun main(args: Array<String>) = Xml2Kotlin().main(args)
 
@@ -108,21 +100,9 @@ class Xml2Kotlin : CliktCommand() {
         help = "Override Kotlin package for generated files"
     )
 
-    @OptIn(ExperimentalXmlUtilApi::class)
     override fun run() {
         val packageOverride = outputPackage?.trim()?.takeUnless { it.isEmpty() }
-        val xml = XML {
-            isUnchecked = true
-            policy = object : DefaultXmlSerializationPolicy(policy) {
-                override fun handleUnknownContentRecovering(
-                    input: XmlReader,
-                    inputKind: InputKind,
-                    descriptor: XmlDescriptor,
-                    name: QName?,
-                    candidates: Collection<Any>
-                ): List<ParsedData<*>> = emptyList()
-            }
-        }.decodeFromString<XmlRootNode>(input.readText())
+        val xml = introspectionXml.decodeFromString<XmlRootNode>(input.readText())
         if (!keep) {
             output.deleteRecursively()
         }
